refactor: decouple HttpApi from cli/commands/daemon
In the past API was exposed via HTTP Server only when jsipfs daemon was run
from the commandline, so src/http/index.js was also responsible for
orchestration that is not related to HTTP itself.

This refactor moves code that is not related to HTTP Servers into
standalone-daemon.js, which is easier to reason about, and unlocks use
of HttpApi in contexts other than commandline jsipfs daemon,
such as Firefox with libdweb or Chromium-based web browser with chrome.sockets APIs.

Refs.
https://github.com/ipfs-shipyard/ipfs-companion/issues/664

License: MIT
Signed-off-by: Marcin Rataj <lidel@lidel.org>